Move the position check
